+++ CHANGELOG.md
@@ -42,6 +42,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.12.0] - 2023-01-10
 
+### Changed
+- Enabled ordertax api testcase in promotional product & in externalseller tests 
+
 ### Added
 Added option to set status for the "AUTHORIZED_RISK_DECLINED" case